+++ google-cloud-spanner/src/main/java/com/google/cloud/spanner/SpannerOptions.java
@@ -99,6 +99,7 @@ public class SpannerOptions extends ServiceOptions<Spanner, SpannerOptions> {
   private final int prefetchChunks;
   private final int numChannels;
   private final String transportChannelExecutorThreadNameFormat;
+  private final String databaseRole;
   private final ImmutableMap<String, String> sessionLabels;
   private final SpannerStubSettings spannerStubSettings;
   private final InstanceAdminStubSettings instanceAdminStubSettings;
@@ -564,6 +565,7 @@ public class SpannerOptions extends ServiceOptions<Spanner, SpannerOptions> {
             ? builder.sessionPoolOptions
             : SessionPoolOptions.newBuilder().build();
     prefetchChunks = builder.prefetchChunks;
+    databaseRole = builder.databaseRole;
     sessionLabels = builder.sessionLabels;
     try {
       spannerStubSettings = builder.spannerStubSettingsBuilder.build();
@@ -674,6 +676,7 @@ public class SpannerOptions extends ServiceOptions<Spanner, SpannerOptions> {
 
     private int prefetchChunks = DEFAULT_PREFETCH_CHUNKS;
     private SessionPoolOptions sessionPoolOptions;
+    private String databaseRole;
     private ImmutableMap<String, String> sessionLabels;
     private SpannerStubSettings.Builder spannerStubSettingsBuilder =
         SpannerStubSettings.newBuilder();
@@ -731,6 +734,7 @@ public class SpannerOptions extends ServiceOptions<Spanner, SpannerOptions> {
           options.transportChannelExecutorThreadNameFormat;
       this.sessionPoolOptions = options.sessionPoolOptions;
       this.prefetchChunks = options.prefetchChunks;
+      this.databaseRole = options.databaseRole;
       this.sessionLabels = options.sessionLabels;
       this.spannerStubSettingsBuilder = options.spannerStubSettings.toBuilder();
       this.instanceAdminStubSettingsBuilder = options.instanceAdminStubSettings.toBuilder();
@@ -829,6 +833,17 @@ public class SpannerOptions extends ServiceOptions<Spanner, SpannerOptions> {
      */
     public Builder setSessionPoolOption(SessionPoolOptions sessionPoolOptions) {
       this.sessionPoolOptions = sessionPoolOptions;
+      return this;
+    }
+
+    /**
+     * Sets the database role that should be used for connections that are created by this instance.
+     * The database role that is used determines the access permissions that a connection has. This
+     * can for example be used to create connections that are only permitted to access certain
+     * tables.
+     */
+    public Builder setDatabaseRole(String databaseRole) {
+      this.databaseRole = databaseRole;
       return this;
     }
 
@@ -1218,6 +1233,10 @@ public class SpannerOptions extends ServiceOptions<Spanner, SpannerOptions> {
 
   public SessionPoolOptions getSessionPoolOptions() {
     return sessionPoolOptions;
+  }
+
+  public String getDatabaseRole() {
+    return databaseRole;
   }
 
   public Map<String, String> getSessionLabels() {